Change to randomising speed in discrete values and angle
Magnitude of velocity (a.k.a. speed) is not randomized across discrete values.
Angle is randmized as well.

Particle bound is also confined to a circle.

Together with discrete speeds this results in a cool converging and divering
pattern. It seems liek the floating number math adds some sligh variability as
the simulation progresses.