Add beat_circle particle pattern generator
Particles travel in a straight line from the center of the circle to the perimeter.
Velocities are set at discrete values that are multiples of each other.
After a time floating point math causes some particles to travel slightly slower
or faster than the initial discrete velocity adding to the appeal.